dom/html/test/imports/test_CSP_sandbox.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1106713
-->
<head>
  <title>Test for Bug 1106713</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1106713">Mozilla Bug 1106713</a>
  <script type="text/javascript">
    SimpleTest.waitForExplicitFinish();
    function go() {
      var ifr = document.getElementById('iframe1').contentWindow;
      ok(!ifr.scriptExecuted, "script is not allowed to run");
      SimpleTest.finish();
    }

  </script>
  <iframe src='file_CSP_sandbox.html' sandbox="allow-same-origin" onload="go()" id="iframe1"></iframe>
</body>
</html>
